requirements.txt
# Providers and libraries used by the legacy scheduler migration example DAGs.
airflow-blueprint>=0.2.0
apache-airflow-providers-databricks
apache-airflow-providers-http
apache-airflow-providers-sftp
apache-airflow-providers-snowflake
apache-airflow-providers-ssh
